Concise commit message:
```
refactor: update YukiHookAPI entry point and optimize canvas math

- Update `GenesisHookEntry` to use `class` instead of `object` and initialize via `configs` for `isDebug` state.
- Optimize coordinate calculations in `CanvasViewModel` by using `Float` overloads for `sin` and `cos`, removing unnecessary `Double` conversions.
- Add `PERPLEXITY` to the `AgentType` enumeration.
- Update `GenesisHookEntry` documentation to reflect the correct KSP-generated entry class name.
```

+++ aura/reactivedesign/collabcanvas/src/main/kotlin/collabcanvas/ui/CanvasViewModel.kt
@@ -114,8 +114,8 @@ class CanvasViewModel @Inject constructor(
             while (true) {
                 if (isConnected) {
                     t += 0.05f
-                    val auraX = 250f + 150f * sin(t.toDouble()).toFloat()
-                    val auraY = 350f + 100f * cos((t * 0.7).toDouble()).toFloat()
+                    val auraX = 250f + 150f * sin(t)
+                    val auraY = 350f + 100f * cos(t * 0.7f)
 
                     webSocketService.sendCursorUpdate(
                         canvasId = roomId,
